refs #30 do not use default filter values in subscription form
The form binding code in formFactory uses getValue on filters.
We now pass $use_default_values=false to getValue so that it doesn't use
default filter values if the value is null, but rather keeps it null

This fixes a bug where a subscription with no distrelease at all in the
subscription_elements would be displayed as if it had one, the default
one. Now it is correctly displayed as "All".